requirements.txt
dearpygui
requests
pygame-ce